Started implementing an empirical model class.
An empirical model is a relatively straight forwars model class that is
essentially based on the following: the probability of an event is the
emperical probability that such an event occurs.

It's implementation is mostly based on already existing method/code for
data operations. To clean it up I refactored some functionality into its
own methods and put htem in data_operations.py (density(),
probability()).

Moreover a bigger problem surfaced: The generic way I deal with conditioning of
models is not correct. At the moment conditioning is only really applied
when a dimension is finally removed. However, it should be applied at
the time of conditioning. The degree of fatality of this error is still to be
determined..