Make probabilistic model class more flexible
The part of loading training data and creating a probabilistic model is
moved to the 'if name == __main__' part of the script. This means that
theoretically one can now specify any probabilistic model, then give it
to the class constructor as a parameter and then it can be visualized in
lumen.