Fix bug in test_conditionout
The variables in samples do not have the same order as the variables in fields.
Since this was not noticed before, I attempted to address the same variable in
samples and in fields by using the same index. This index however addressed
different variables, leading to the test failing, where it should pass. Variables
are now addressed by name instead of index to avoid this.