introduce the evidence parameter to Model.predict
With this parameter the caller may provide data that is used as conditions for modelling.
The idea is, for example, that it is easily possible to predict a target dimension for a given set of observations.

I am very pretty sure that the current implementation will not work as I have not even tried running it.