Set default value for split_data argument
The split_data argument was introduced to avoid the splitting into test
and training data, if needed. I forgot to set a default value for it, so
that each model class that did not specify split_data threw an error. This is fixed now by setting split_data to True by default