Set extent not for independent variables
For the goal of including independent variables in the visualization of the lumen frontend,
these variables have to be kept in fields. Fields and samples, however, should contain the
same variables in the same order (I think because they are adressed somewhere by the
same index and things get mixed up if the order is different).
Therefore, samples als has to contain independent variables. By definition, no samples are
created for independent variables, so the corresponding columns in samples contain only NaNs.
adopt_all_extents() cannot deal with columns where only NaNs occur, so I implement a check
that ensures that the extent is only set for columns where NOT all elements are NaN