Add independent variables to samples
For the goal of including independent variables in the lumen frontend, those variables
have to be kept in fields. Since fields and samples should contain the same variables in
the same order, samples also has to contain those variables (If the order is different,
things get mixed up , I think because the same index is used to adress fields and samples
somewhere). This is achieved by adding a column to samples for each variable in data.
The columns consists solely of NaNs since by definition no posterior samples are drawn for
independent variables