Check in _density method if samples consist of only NAs
For the display of marginal distributions in the lumen frontend, densities
are queried for just one variable. If this variable is independent, its samples
consist only of NaNs which causes the _density ,ethod to crash. To prevent
that, a check is implemented and  NaN is returned  if all samples consist
of NaNs