Fix bug where ppc samples cannot be drawn when no shared vars are given
Previously it was assumed that data and samples have the same length. But
this is only the case when the data is changed by using shared variables.
To fix this, it is checked if shared vars are present. If not, ppc
samples are selected a different way: For each sequence of points
belonging to one parameter set, only the first point is selected. This
is problematic if we have independent variables, since only the first
of these is considered in each selected point. If we have indepedenten
variables, they should be encoded as shared variables however, meaning
that this way of selecting sample points is not applied then. Encoding
independent variables as shared variables has yet to be enforced