Adjust get_numerical_fields for shape parameter
Previously, fields were updated by calling the corresponding method
with the samples dataframe and the variable names of the model.
When having multiple vars within one var via the shape parameter,
the column names of the samples df are different than the var names
of the model. Therefore, the method is now called with the
column names of the samples df instead of the variable names of
the model.