Make the number of samples dependent on the length of the data
Initially, the number of samples drawn from the posterior distributions
was set to a fixed number. However, there are models that require their
data to have a fixed length, so when the number of posterior
samples did not match the required number of the model, there was
a conflict. As a first workaround, the number of posterior
samples is now always equal to the number of observations in the
data. That is not ideal: For one, the number of posterior samples
for the model parameter is now also equal to the length of
observations, which is not necessary. Also the number of
observations in the data migth be far lower than the desired
number of posterior samples. So things have to be further
improved here in the future