Adjust number of posterior samples
Previously, the number of posterior samples was set to the length
of the input data. This was not ideal since the length of the
input data might be very low. Posterior samples with a length
equal to the input data's length are now generated multiple
times, until a prespecified number of samples is reached. The
implementation is not yet finished: By now, it works only if
the prespecified number of samples is a full multiple of the
length of the input data