Revert commit 309f2e
Previously the adopt_all_extents method was changed for the PyMC3
model class so that no extent was set if the samples were all NAs.
This is not necessary anymore since the PyMC3 samples will now
never include NAs. Since the code piece is therefore redundant
and also leads to errors in non-PyMC3 models, it is removed now