Rename aggregation method
In the aggregate_data()-method in data_aggregation, the operations
'maximum' and 'average' are computed in the same way. In my
understanding, the arithmethic mean is computed in both cases.
Isn't this misleading for the maximum? In my implementation for
the KDEModel class, I assume that a maximum of the KDE is the same
as the maximum of the data and so I use the aggregate_data()-method
in data_aggregation. Since the maximum there is actually the
arithmethic mean, I renamed my method from '_maximum' to
'_arithmetic_mean'.