Change _conditionout method
The previous implementation of the method did not work since it
required the data of the dimensions in remove. This data is
however marginalized out before by the _marginalize() method.
As a workaround I recreated the approach from the EmpiricalModel
class, where there is an additional data object that stores
the same data as self.data during the fitting process. This
object is then not touched by _marginalize() and can be
conditioned on later. At the end of the _conditionout() method,
only those indexes in self.data are kept that appear also in
the additional data object