Add finding of maximum density by numeric optimization
Finding the point of maximum density of a model by numerically
getting the maximum of the density function is difficult
bevause the density function is not necessarily convex. I
tried to get around it by selecting multiple starting points
and running the optimization for each point, then selecting
the point with the highest density among them. This seems
to work, but is extremely slow and not acceptable for
even medium sized models. A faster method has to be implemented
in the future.