Change starting point for optimization in maximum
For the calculation of the point of maximum probability density,
an optimization problem on the density function is solved.
Previously, the starting point for this optimization was
at the zero-vector. It is now changed to the mean of the extent
for each variable. This is still not ideal since there is no
guarantee that the global optimum is found, but hopefully
better than the previous starting point