Change calculation of maximum
Instead of creating a grid of possible starting points and solving
an optimization problem for each of those points, I now just
take the mean of the samples as single  starting point. This
probably makes the method more error-prone, but much faster. It
is not ideal since there is no guarantee that the global optimum
will be found