Work on integration of categorical variables in kde
Work on support of models with both continuous and categorical
variables. Getting a density in those cases is achieved by
calculating the conditional probability density for the
numerical variables conditioned on the values of the
categorical variables, as well as the marginal probability
density for the values of the categorical variables. Both densities
can then be multiplied to get the joint probability density
of numerical and categorical variables:
p(num,cat) = p(num|cat)*p(cat)