Implement Calculation of conditional probability density
To get a density for models with numerical as well as catgorical
variables, the conditional density for the numerical values
conditioned on the categorical ones has to be computed. For that,
I condition the model data on the categorical values of the
point for which the density is to be calculated. That in turn
requires this point to have the same order of variables as the input
data, since I loop through the variables per index. So the order
of the variables of the point is transformed the same way as
the model data was transformed on input