Implement maximum method for categorical and numerical vars
The point of maximum probability is computed by solving an
optimization problem for each possible value of the
categorical variables.  The maximum of a conditional
probability for the numeric variables, conditioned
on the categorical values is computed this way. To get the
joint probability of numerical and categorical variables,
the conditional probability then is multiplied by
the marginal probability of the chosen categorical value,
that is, its relative frequency in the data. The point
that has the maximum probability from all the combinations is
then chosen as the global maximum.

The method works also if there are no categorical variables in
the model