Keep kernel density estimators
Previously, a new kernel density estimator was generated each time
the density was queried. Now, the first time a kernel desnity
estimator for a specific value of the categorical variables is
generated, it is stored in a dictionary. So the next time
this specific kernel density estimator is needed, it does not
need to be generated again but can be taken from the dictionary