Debug generation of posterior predictive samples
Previously, an index was used with a length equal to the number of the
original data. This lead to errors when the actual unique ppc samples
had a lesser length that was only expanded later during the grid
generation to the length of the original data. The length of
the index is changed now to the length of unique values in the ppc