maximum: Return NaN if func value is NaN
When all the values of the density function return nan, the maximum
method previously returned its starting point as the maximum.
However, if the density is alway nan, it makes more sense not
to return anything as the maximum. SO it is changed now to return
NaN as a maximum in these cases.