Account for differences in specified and actually calculated number of posterior samples
The number of posterior samples to generate is specified in the class
instance attributes. The actual number of generated samples is derived
by taking the n-th root of this number, where n is the number of
independent variables in the model, then generating a number of samples
equal to the result for each independent variable, and then performing
a cross join of all the independent variables, thus expanding the
number of samples back to the number specified in the class instance
attribute. However, since taking the n-th root as an integer
introduces rounding errors, the final number is in most cases
lower than the originally specified number That's why the difference
between the actual and the desired number of samples is filled up
by repeating the x first rows of the dataframe, with x being the
difference between the desired and the actual number