Include independent variables in sample method
The _sample() method so far only sampled from the random variables in the
model. This is of course correct, we don't want to sample from
covariates. However, the higher-level sample() method requires data from all
variables in the model. So when the model has independetn variables, the
_sample() method has to return also data for these variables. This is now implemented by
basing all the samples on the first value of the independent variables
and returning this first value for the independent variables.