Fix independent variable generation when number of variables is big
When there is a big number of independent variables, the number of unique values
generated for each of these variables becomes very small. If this number
becomes 1, building the cartesian product of all the unique values results
in only one row, which is not enough. Workaround: Repeat the appending
of rows until the desired number of rows is reached