Catch case without independent variables in sampling
When generating samples for observed dependent variables, the
according pymc3 function automatically generates a number
of samples equally to the length of the observed data.
When there are no independent variables, I cannot control
this number so I can't rely on it being consistent with the
rest of the code. Therefore I just take the first point of each
sequence, which is not wrong since this is only done when there
are no independent variables