Fix bug in copy method
The bug was that during copy(), the shared_vars attribute of the copy-model
was initially set to an empty dict, even when there were no shared
variables in the model. This lead to errors when it was expected that
shared vars did not exist, but instead an empty dict existed. It is now
fixed so that shared_vars is not set at all if there are no shared
vars present