Do not change shared vars during _marginalizeout() anymore
shared variables are now the same between all copies of a model.
Permanent changes of the shared variables thus have to be
prevented, since these changes would affect all copies. So
the shared variables are not included in the marginalize process
anymore. TODO: Deal with the errors that are caused by this change