Rework generation of samples for independent variables
During the sampling process every variable of shared_vars has to be
set to new values. If only the variables of the current model
were set, there could be disvrepancies between the variables
of the current model and the original model. So samples
for EACH independent variables are generated, but only
those variables that belong to the current model are propagated
to the final result.