Write comments
Following comments are written:
Note: The shared_vars attribute is not copied. Rather, the same shared_vars
object is used for EVERY copy of the model. Copying the object would be
useless since the model_structure is linked with the original object
nevertheless. This means that the shared vars attribute must not be
changed permanently, because doing so would propagate to all model copies