Put data into original order after set_data_mixed
I think that pm.sample() uses the original data values. When I change or
shuffle the values of the independent variables, this leads to incorrect
sampling results, since the data for the  dependent variables cannot be
changed (at least not with the current implementation). The data attribute
however is shuffled during set_data_mixed. So I put the data back in the
original order after set_data_mixed