Account for case when domain is singular
During conditioning, minimum and maximum of a field's domain are
extracted using the domain.value attribute. Normally this attribute
is a list of two values, min and max. However, if min and max are the
same, only this one value is stored in the attribute, leading to
an error when it is tried to index the list. This is fixed now
by checking which case occurs beforehand, and extracting the
values accordingly