Fix number of posterior samples to length of data
Previously, the number of posterior samples could be set arbitrarily.
However, I noticed that this leads to a problem in models where
dependent and independent variables interact with each other, for
example when the sum of an independent and a dependent variable
is computed. When the number of posterior samples differs from
the number of data points, independent and dependent variables
have different lengths, so the abovementioned operations cannot
be computed and throw an error. This happens during the generation
of posterior predictive samples. To fix that, the number of
posterior samples are now fixed to the length of the data.