fill data namespace package
This complex commit moves all of the freely available data from the
private repository that was used so far to this namespace package.
Having a centrally and independently available data repository helps to
organize code, tests, jupyter notebooks and more.

Consequently a large number of files had to be adapted due to the
different locations and import structure.